pyproject.toml
[tool.poetry]
name = "project on fastapi"
version = "0.1.0"
package-mode = false
description = "mini project FastAPI"
authors = ["Vlad <vlad111121562@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.6.4"}
libgravatar = "^1.0.4"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
